requirements.txt
python-dotenv
openai
paddleocr
paddlepaddle
pandas
rapidfuzz